requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
authlib==1.3.0
itsdangerous==2.1.2
cryptography==42.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
